Separated assets publication into its own command
Refactored InstallCommand to better check if a directory exists
